requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.5
numpy>=1.26
pandas>=2.0
pyyaml>=6.0
duckdb>=1.0.0
